requirements.txt
# Core FastAPI + servidor
fastapi==0.110.2
uvicorn[standard]==0.30.1
orjson==3.10.7

# RAG
chromadb==0.5.5
sentence-transformers==2.7.0
transformers==4.44.2
torch==2.4.0
pypdf==4.2.0

# LLM remoto
groq==0.11.0
httpx==0.27.2

# Utilidades
python-dotenv==1.0.1
